requirements.txt
alembic==1.16.4
black==25.9.0
openpyxl==3.1.5
pydantic==2.11.7
python-telegram-bot==22.4
SQLAlchemy==2.0.43
